+++ CHANGELOG.md
@@ -9,13 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Check for empty transient value when fetching term language.
 - Constant `UBB_SETTINGS_READONLY` for read only settings in the backoffice settings page.
 - Internal option `ubb_settings_manual_changes` for keeping track if manual changes have been made to the settings.
 
 ### Changed
 
+- Improved get post language handling of empty values.
 - Options in options page are now modifiable from the values set in the `ubb_options` filter.
 - API rest url when Unbabble is set to directory routing no longer has the language directory applied.
+
+### Fixed
+
+- Term language set when term has no language.
+- Missing post type in url when creating a term translation to stay in the same post type Menu.
 
 ## [0.4.8] - 2024-07-30